backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
tenacity==9.0.0
python-dateutil==2.9.0
orjson==3.10.13
python-dotenv==1.0.1
structlog==24.4.0
numpy==2.2.1
PyJWT==2.10.1
